requirements.txt
python-dotenv==1.0.1
pandas>=1.3.0
requests>=2.25.0